Fixed various eclipse warnings. Note: VirtualSocketFactory.createServerSocket(..., Map properties) is changed to (..., Properties properties). It is not correct to address a properties object as a map.
git-svn-id: http://gforge.cs.vu.nl/svn/ibis/smartsockets/trunk@10939 aaf88347-d911-0410-b711-e54d386773bb

+++ touchgraph/src/com/touchgraph/graphlayout/interaction/HyperScroll.java
@@ -173,7 +173,7 @@ public class HyperScroll implements GraphListener {
         double x2 = inverseArray[i];
         double x1 = inverseArray[i - 1];
         double j = (dist - x1) / (x2 - x1);
-        return (((double) i + j - 1) / 200.0 * width);
+        return ((i + j - 1) / 200.0 * width);
     }
 
     class HyperLens extends TGAbstractLens {
